flash

+++ Assets/_Scripts/Enemy/EnemyInform.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 [Serializable]
 public class EnemyInform
 {
-    public int enemyHp;
-    public int ememyDamage;
+    public float enemyHp;
+    public float ememyDamage;
     public float ememySpeed;
 }
